+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 7.3.2 under development
 
-- no changes in this release.
+- New #144: Introduce `WebViewRenderer` instead of `ViewRenderer` which is marked as deprecated (@vjik)
 
 ## 7.3.1 December 19, 2025